+++ tests/test_sloan_log.py
@@ -33,6 +33,7 @@ class TestSloanLog(unittest.TestCase):
         args.morning = False
         args.verbose = True
         args.noprogress = False
+        args.mirrors = True
 
         ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
         b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
@@ -68,6 +69,7 @@ class TestSloanLog(unittest.TestCase):
         args.log_support = False
         args.morning = False
         args.verbose = True
+        args.mirrors = False
 
         ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
         b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
@@ -99,6 +101,7 @@ class TestSloanLog(unittest.TestCase):
         args.log_support = False
         args.morning = False
         args.verbose = True
+        args.morrors = False
 
         ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
         b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
@@ -128,6 +131,7 @@ class TestSloanLog(unittest.TestCase):
         args.log_support = True
         args.morning = False
         args.verbose = True
+        args.mirrors = False
 
         ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
         b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
@@ -157,6 +161,7 @@ class TestSloanLog(unittest.TestCase):
         args.log_support = True
         args.morning = False
         args.verbose = True
+        args.mirrors = False
 
         ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
         b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
@@ -167,6 +172,38 @@ class TestSloanLog(unittest.TestCase):
         log.sort()
         log.count_dithers()
         log.log_support()
+
+    def test_mirror_numbers(self):
+        """Runs on an old dataset that I know used to run successfully"""
+
+        class Dummy(object):
+            pass
+
+        args = Dummy()
+        args.sjd = sjd.sjd()
+        args.print = False
+        args.summary = False
+        args.data = False
+        args.boss = False
+        args.apogee = False
+        args.p_boss = False
+        args.p_apogee = False
+        args.log_support = False
+        args.morning = False
+        args.verbose = True
+        args.mirrors = True
+
+        ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
+        b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+        ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
+        b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
+        log = sloan_log.Logging(ap_images, b_images, args)
+        log.parse_images()
+        log.sort()
+        log.count_dithers()
+        log.log_support()
+        log.mirror_numbers()
+
 
 if __name__ == '__main__':
     unittest.main()